Fix implementation errors in the controllers. Whole system works
Remove the sign change of the errors matrix, as this is already
done in the kalman node.

Invert the order of the subtraction operation for getting the
errors.

Reduce the Kp to 0.02. The speeds evolution is very slow, so the
Kp has to be low too, in order to adapt accordingly to the system.

Add uniform error to the starting position of the drones